Using di.xml to fix installation problems
When running `setup:install`, all console commands are loaded. If a command relies on a class(es) that have database calls in their constructor, the installation will error out.

This is a scenario that is not often tested as this is an installation from scratch with the Composer dependencies loaded.

This update has been working well for us.