fix error an empty cron expression.
If the cron expression is empty, the schedule is not created.

There are cases where the cron expression is empty. Extensions like Mageplaza Cron Scheduler and Magefan Cron Scheduler have a feature to disable cron jobs. To disable a specific job code, the cron expression is left empty. However, this causes an issue where no other cron schedules are generated because the disabled cron jobs, lacking an expression, trigger errors by m2-ce-cron extention. so I've modified to skip the disabled cron jobs.